[TASK] Add `RuleContainerTest` trait, use in `RuleSetTest`
The trait provides tests for classes implementing `RuleContainer`.
The test methods and data providers have been moved verbatim to the trait from
`RuleSetTest`.

Will be needed for #1194.
